+++ .gitignore
@@ -1,0 +1,4 @@
+
+.idea/
+__pycache__/
+*.egg-info/
