+++ cvd_vae/models.py
